% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemcalculator.R
\name{compute_mass}
\alias{compute_mass}
\title{Computes the molar mass of the given chemical compound.}
\usage{
compute_mass(chemical)
}
\arguments{
\item{chemical}{The molecular formula given as a string.}
}
\value{
The molar mass of the given formula in g/mol.
}
\description{
Computes the molar mass of the given chemical compound.
}
\examples{
\dontrun{
compute_mass("H2O") ## returns 18.01528

compute_mass("C12H22O11") ## returns 342.3

compute_mass("Al2(SO4)3") ## returns 342.15

compute_mass("(NH4)HS") ## returns 51.107
}
}
