% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemcalculator.R
\name{.check_chemical_format}
\alias{.check_chemical_format}
\title{Check that the chemical formula has correct format}
\usage{
.check_chemical_format(chemical)
}
\arguments{
\item{chemical}{chemical formula to check}
}
\value{
True or FALSE for if chemical formula has the correct format
}
\description{
Check that the chemical formula has correct format
}
\examples{
.check_chemical_format("H2O")          ## True
.check_chemical_format("NaaaaaaaaaOH") ## False
}
