% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemcalculator.R
\name{.chemical_elements}
\alias{.chemical_elements}
\title{Decomposes a chemical to it's elements.}
\usage{
.chemical_elements(chemical)
}
\arguments{
\item{chemical}{The molecular formula of the given chemical compound given as a string.}
}
\value{
Dataframe of the chemicals elemental components and their counts.
}
\description{
Decomposes a chemical to it's elements.
}
