% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemcalculator.R
\name{percent_mass}
\alias{percent_mass}
\title{Calculates the percentage mass of an element (or compound) in a compound}
\usage{
percent_mass(compound, element)
}
\arguments{
\item{compound}{the chemical formula of the full compound}

\item{element}{the chemical formula of the element or compound of interest}
}
\value{
the percentage mass of the element or compound of interest
}
\description{
Calculates the percentage mass of an element (or compound) in a compound
}
\examples{
\dontrun{
percent_mass("H2O", "O") ## returns 88.79
percent_mass("H2O", "H") ## returns 11.19
percent_mass("NaOH", "OH") ## returns 42.52
}
}
